Simplify inheritance
