+++ libs/common/src/abstractions/i18n.service.ts
@@ -1,11 +1,7 @@
 import { Observable } from "rxjs";
 
-export abstract class I18nService {
+import { LocaleService } from "./locale.service";
+
+export abstract class I18nService extends LocaleService {
   locale$: Observable<string>;
-  supportedTranslationLocales: string[];
-  translationLocale: string;
-  collator: Intl.Collator;
-  localeNames: Map<string, string>;
-  t: (id: string, p1?: string | number, p2?: string | number, p3?: string | number) => string;
-  translate: (id: string, p1?: string, p2?: string, p3?: string) => string;
 }
